pyproject.toml
[project]
name = "crewai-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "cohere>=5.17.0",
    "crewai>=0.177.0",
    "crewai-tools>=0.0.1",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "langchain-cohere>=0.4.5",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "python-dotenv>=1.1.1",
    "serpapi>=0.1.5",
    "uvicorn>=0.35.0",
]
